refactoring of hunter-app.js
